pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-agent-core"
version = "0.1.0"
description = "Reusable database-agnostic NL2SQL runtime with LlamaIndex orchestration"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "data-agent-core contributors" }]
dependencies = [
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "sqlalchemy>=2.0",
  "typer>=0.12",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "llama-index-core>=0.11.0",
  "llama-index-llms-google-genai>=0.2.0",
  "duckdb>=1.0",
  "duckdb-engine>=0.13",
  "pyarrow>=16.0",
  "mcp>=1.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
  "pre-commit>=3.7",
]
postgres = ["psycopg[binary]>=3.2"]

[project.scripts]
data-agent = "data_agent_core.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
mypy_path = "src"
